Add support for 18 character organization IDs
The regex in the `AccessToken::getOrgId()` method first attempted to
match the Org prefix followed by 12 OR 15 word characters. This resulted
in org IDs of 18 characters being stripped to 15 characters in length.

The fix for this was to reverse the match - first try and match the Org
prefix followed by 15 characters *then* 12 characters for older org IDs.